nexum_ai/requirements-lock.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements-lock.txt requirements.txt
#
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via tqdm
diskcache==5.6.3
    # via llama-cpp-python
filelock==3.20.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.12.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.36.0
    # via
    #   -r requirements.txt
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via requests
jinja2==3.1.6
    # via
    #   llama-cpp-python
    #   torch
joblib==1.5.2
    # via scikit-learn
llama-cpp-python==0.3.16
    # via -r requirements.txt
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.3.5
    # via
    #   -r requirements.txt
    #   llama-cpp-python
    #   scikit-learn
    #   scipy
    #   transformers
packaging==25.0
    # via
    #   huggingface-hub
    #   transformers
pillow==12.0.0
    # via sentence-transformers
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   transformers
regex==2025.11.3
    # via transformers
requests==2.32.5
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.7.0
    # via transformers
scikit-learn==1.7.2
    # via
    #   -r requirements.txt
    #   sentence-transformers
scipy==1.16.3
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.1.2
    # via -r requirements.txt
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.1
    # via transformers
torch==2.9.1
    # via
    #   -r requirements.txt
    #   sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.57.3
    # via sentence-transformers
typing-extensions==4.15.0
    # via
    #   huggingface-hub
    #   llama-cpp-python
    #   sentence-transformers
    #   torch
urllib3==2.6.1
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
